hv: add PCI PIO BAR reprogramming support
Linux guests will reprogram PCI BARs in some cases. A typical scenario
is when a new device is hotplugged. Before this patch, hypervisor forces
PIO bar passthru and guest will got a #GP when it reprograms a new BAR
base.

This patch introduces a trap and relay method to handle accesses to PIO
BARs for passthru PCI devices.

Signed-off-by: Qiang Zhang <qiang4.zhang@linux.intel.com>