DM: vhost-user protocol support
We implement the vhost-user protocol, which can transfer the
dataplane to vhost-user backend, such as virtiofsd daemon.

The vhost-user protocol is similar to the vhost backend which
transfers the dataplane in kernel.

The vhost/vhost-user backend will take over the virtqueues,
receive virtqueue's kick/call events through eventfd, and process
msgs in virtqueues.

Because the similarity between vhost and vhost-user,
The code implementation is also similar.

Tracked-On: AUTOMOTIVE-1232
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>