dm: virtio-fs: add the multiple request queues support
In current implementation, virtio-fs only supports 2 queues, including
1 high-priority queue and 1 request queue.
Due to the lock contention in the FE driver, only one request queue is not
optimal from the performance perspective.

In order to improve the performance, this patch adds the multiple request
queues support in virtio-fs BE.
 - By default, the number of queues is 2, including 1 high-priority queue and
   1 request queue.
 - Example to specify multiple queues:
   `add_virtual_device   15    virtio-fs tag=tagsq,socket=/data/sckt,num_queues=8`
   In this example, the total number of queues is 8, including
   1 high-priority queue and 7 request queues.

Note:
This patch has the dependency of virtiofsd and virtio-fs FE driver.
It can only work if multiple request queues are supported in all components,
including acrn-dm, virtiofsd and virtio-fs FE driver.

Tracked-On: ACRN-10683
Signed-off-by: Shiqing Gao <shiqing.gao@intel.com>
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>