misc: add ivshmem-timesync tool
The idea of ivshmem timesync is to send time sync msg between VMs via
ivshmem. As all VMs have the same TSC counter, so the differ value
between TSC and systime can be used to sync the systime very accurately.

The major uncertainty is introduced when obtaining the differ value on
the sync master VM, as it involves syscall and cannot avoid context
switch. The total time of obtaining differ must be recorded, and large
values should be dropped.

The ivshmem-timesync tool is used to demostrate how to sync time between
VMs via ivshmem, and evaluate the accuracy of this apoach.

Tracked-On: ACRN-10199
Signed-off-by: Jian Jun Chen <jian.jun.chen@intel.com>
Signed-off-by: Wu Zhou <wu.zhou@intel.com>