dm: virtio-spi: add spi device event notification
SPI Bus is a single master bus that all transfers are started by the SPI
Master. In many use cases, SPI devices will leverage a side band signal like
GPIO to notify the master that it has data to be processed.

In the virtualization environment, we provide a SPI device notification
mechanism in the Virtio SPI controller with a dedicated event queue like
that of Virtio GPIO.

Guest pushes an IRQ enable/unmask request for a chip select to the event
queue to enable the notification for the SPI device. When SPI device BE
decides to notify the FE SPI device driver, the IRQ request descriptor is
"used" and return back to the FE. At the same time, the IRQ is disabled. FE
needs to push the IRQ request again for next notification.

Since it's the SPI device BE to decide when to inject a event to the FE,
we proivde an interface for device backend drivers and a event proxy thread
for external device emulators to inject notification to virtio SPI frontend.

Tracked-On: AUTOMOTIVE-1238
Signed-off-by: Qiang Zhang <qiang4.zhang@intel.com>