DM: virtio-fs: support multi virtio-fs instances
We propose this patch to enable multiple virtio-fs
instances in a VM.
Due to the design of the virtio-fs daemon, which is
accepting a socket connection only once and closing
the socket in acrn-dm will cause the daemon to exit.

The method to support multiple instances is that
we maintain a global array, only link once for each
virtio-fs slot, and do not execute close socket_fd.

Let the acrn-dm automatically release the socket_fds
when it exits. In this way, whether it is a
PCI hot-plug or a VM reboot, a virtio-fs slot will
only be bound to one virtiofs daemon.

Tracked-On: AUTOMOTIVE-1232
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>