dm: disk-logger memory copy optimization
currently acrn-dm use the disk logger to
store log msg under /var/log/acrn-dm/* .

It can be improved in the following aspects:
1. the length of one log msg is limited.
Which is the DISK_LOG_MAX_LEN, when the msg is
longer than that, will be truncated.
2. memory copy when writting to log files.
In write_to_disk(), will copy the log msg
to buffer[], than write to file. Which is
not very efficient

Make changes for disk logger,
1. cancel the log msg length limit.
2. delete the memory copy when writting to file.

Tracked-On: AUTOMOTIVE-1494
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>