misc/hv: mca: Add new VM conf: mcbanks_bitmask
This patch adds a new per-VM config mcbanks_bitmask.

This bitmask represents the MCi_xx banks to be passed through.
Defaults to zero. For example, if we'd like to passthrough MC banks 0-4
to a guest, set <mcbanks_bitmask>0xf</mcbanks_bitmask> for this VM.

We have the following assumptions for this bitmask:
1. 1s are continuous
2. 1s starts from position 0
3. Hardware MC banks are organized in the following order:
   per-thread, per-core, per-package

If this bitmask is zero, MC bank passthrough and #MC/CMCI injection is
disabled for this guest.

It is the config-tool/user's responsbility to make sure that
1. Guest who sets this bitmask owns all pCPUs assigned to it
2. Per-core or per-package banks are not passed-through to guest that
doesn't own an entire core/package.

Tracked-On: ACRN-10312
Signed-off-by: Yifan Liu <yifan1.liu@intel.com>