dm: add the unmap memseg interface
dm may do memseg map for some pci bars during the
device init phase. Which is used to present some
bar-mem to Guest VM.

But when deinit the device, dm will not unmap these
bar mem.                                               

Here is to add the unmap memseg interface, so that
the device can call this interface to unmap memseg.

Note: this interface also depends on the HSM driver
patch.
can refer:
https://github.com/intel-innersource/virtualization.hypervisors.acrn.acrn-automotive.acrn-kernel/pull/28/files

Tracked-On: ACRN-10837
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>