dm: add a lock for logfile write
acrn-dm writes some content to a logfile, using the function
fprintf() which is thread-safe. But we also use some
data, such as cur_log_size. It is not thread-safe.

Moreover, relying on cur_log_size to create a new logfile
and other operations are not thread-safe.
Therefore, we need to add a lock to implement a thread-safe
log write().

Tracked-On: ACRN-11213
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>