dm: handling the error of pci_dev_init for VM
In the original design and implementation, the failure of
init a pci device will cause the VM to fail to start.

We want to modify the logic here, which is the failure of
a pci devices will not prevent the VM from booting.

The configuring of a pci device usually is
-s <slot>[:<func>],<emul>[,<config>]
Which will not impact the VM starts when init failed.

We add '-S' here to indicate additional
attributes of the pci device, e.g. it will prevent VM
starting when it inits failed.
-S <slot>[:<func>],<emul>[,<config>]

Tracked-On: AUTOMOTIVE-1501
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>