dm: handling the error of pci_dev_init for VM
In the original design and implementation, the failure of
init a pci device will cause the VM to fail to start.

We want to modify the logic here, and accept the additional
attributes of the pci device so that the failure of
a pci devices will not prevent the VM from booting.

The configuring of a pci device before were
<slot>[:<func>],<emul>[,<config>]

We add an optional flag here to indicate additional
attributes of the pci device, e.g. whether it is necessary
for this VM.

The proposed configuring is
<slot>[:<func>],<emul>[,<dev_flag>,<config>]
e.g. -s, 3,virtio-blk,dev_flag=1,/mnt/vm.img

When parsing the configuring, backward compatibility
will be maintained.

Tracked-On: AUTOMOTIVE-1501
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>