misc: add tpm2-enabled checking logic
There is no more than one vm be assigned to TPM2 enabled.
We supposed there was only one prelaunched vm be set TPM2-enabled
but actually we should tackle more generally cases.
So I calculated the number of vm which is set TPM2-enabled,
if there was more vms we let users know and to remove others.

Tracked-On: ACRN-11381
Signed-off-by: dongpingx <dongpingx.wu@intel.com>