dm: gpio: add mock gpio line emulator
This is a fully emulated gpio line backend. Line state is maintained
internally. However other ACRN DM modules are allowed to change lines
states through dedicated API when the lines are in input state.

Hence this works as a basic facility to emulate complex devices with
gpio as sideband signals.

Tracked-On: ACRN-11476
Signed-off-by: Qiang Zhang <qiang4.zhang@intel.com>