misc: add mandatory_pci functions for script_template
Currently, we add an interface for acrn-dm that accepts -S,
which is used to indicate that this pci_slot will prevent
VM from starting up if it inits failed.

In the launch script, we add corresponding functions,
which are named add_mandatory_passthrough_device and
add_mandatory_virtual_device.
These functions mainly replace -s with -S

Tracked-On: AUTOMOTIVE-1501
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>