hv: Makefile: Change HV_COMMIT_DIRTY variable generation logic
This patch changes the HV_COMMIT_DIRTY to track the uncommitted
diff into HV_COMMIT_DIRTY string.

In build systems such as Yocto, the *offline patches* are applied but no
new commits are created, resulting all offline changes to be placed in
staging build area as "uncommitted changes".

The requirement is: if uncommitted changes exist at build time, we need
to have some methods to compare if two ACRN binaries are built from
exactly the same source code, from the version string alone.

Hence the introduction of the diff hash.

This diff hash is NOT supposed to track the exact content of uncommitted
changes, it is simply used to tell whether the diffs are *different*.

It is expected that
* `git diff` outputs exactly the same content with the same diffs on
different HEADs
* `git diff` outputs different content with different diffs (obvious)
* Build from a non-git-repo (plain folder) should have empty HV_COMMIT_DIRTY
* If command `md5sum` is non-existent, the diff hash is set to
  "unknown".

Tracked-On: ACRN-11694
Signed-off-by: Yifan Liu <yifan1.liu@intel.com>