hv: pcpu bitmap check for VMs
In the previous design, we think that the cpu_affinity of
prelaunch&service VM is already strictly checked, because
it is generated according to the board and scenario xml.

But, when replacing the CPU Die or BIOS closing
hyperthread, resulting in different number of CPUs on the
board, which will cause the VM to fail to start.

Here we re-check the cpu_affinity for VM, and
let the number of used physical CPUs
no more than MAX_PCPU_NUM.

Tracked-On: ACRN-11701
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>