hv: vie: emulate cmpxchg instructions for uc-lock
LOCK prefix is one of 4-group instruction prefixes and it may not be the
first prefix because 4 groups of prefixes can be arranged in arbitrary
order. Thus conceptually, we have to emulate all bus-lock instructions
to support bus-lock(uc/ac) detection.

This patch adds cmpxchg support which may trigger bus-lock if memory is
mapped uncached.

Tracked-On: #8750
Signed-off-by: Qiang Zhang <qiang4.zhang@intel.com>