dm: Introduce plane-display vdisplay back-end
The new plane-display back-end is intended to work with weston, sending
display commands composed of plane info to weston for direct displaying
on hardware planes and bypassing GPU composition.

The back-end connects to weston and sends plane fds for displaying. We
can set the port on which the guest should display by setting port
parameter for virtio-GPU, e.g.,

	`plane-display=$RESOLUTION+0+0:port=DP-1:hz=120`

Then we will use /tmp/proj-DP-1 as socket for communication with weston.
If port is not set, we will fallback to socket /tmp/proj-default.

The new back-end waits for physical vblanks and inject them to guest.
The vblank_flip event added by us includes last flip sequence, we will
use it to ensure flip event would be handed to guest only when flip
happened already.

Tracked-On: ACRN-11929
Signed-off-by: hangliu1 <hang1.liu@linux.intel.com>
Signed-off-by: Xue, Bosheng <bosheng.xue@intel.com>
Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>